Turbo Stream
Was updated the methods in tasks controller to respond to turbo_stream.
By this way, is possible to delete or append/prepend records from the
tasks list in the index view.

Also, was created flash messages with turbo, by this wasy, is possible
to see the flash messages and turbo automatically will updated the
content.

The hello_world javascript controller was renamed with flash and this
controller contains the function to delete a flash message and a
setTimeOut to delete automatically the flash message.